Add files via upload
This program displays a table containing employee names and avg customer satisfaction survey scores for each employee based on records from a .csv file. It also displays the name of the individual with the highest avg. Lastly, the program allows the user to conduct filtered searches within the data records by entering an employee name, and the program displays the employee's overall performance rating (based on avg survey scores) and lists their respective score for each survey on record.